Refactor: 提取共享测试夹具 FakePluginStorage
把 LaunchpadPreferencesTests 里的私有嵌套 FakeStorage 抽成共享
fixture，供 LaunchpadLayoutStoreTests 等复用；加 writeCount 仪表
用于断言「无变化不写盘」（设计风险 R8）。
